requirements.txt
PyQt5-5.13.0